requirements.txt
annotated-types==0.7.0
anyio==4.12.0
arabic-reshaper==3.0.0
asgiref==3.11.0
asn1crypto==1.5.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
cssselect2==0.8.0
dj-database-url==3.0.1
Django==5.2.9
django-filter==25.2
djangorestframework==3.16.1
freetype-py==2.5.1
gunicorn==23.0.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
lxml==6.0.2
Markdown==3.10
oscrypto==1.3.0
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
pycairo==1.29.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
PyMySQL==1.1.2
pypdf==6.4.0
python-bidi==0.6.7
PyYAML==6.0.3
replicate==1.0.7
reportlab==4.4.5
requests==2.32.5
rlPyCairo==0.4.0
six==1.17.0
sqlparse==0.5.4
svglib==1.6.0
tinycss2==1.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
webencodings==0.5.1
whitenoise==6.11.0
xhtml2pdf==0.2.17
